Throw Exception when handle_function fails
When a function is mocked using `Functions::create_function()`, two things are created:

1. The actual function is defined using `eval()`
2. The function handler is bound using `Handler:register_handler()`

When the test is complete and `Functions::flush()` is called, it removes the
handler, but nothing removes the function itself, which will continue to call
`Handler::handle_function()` in later tests, except that this will now be a
noop.

This can create problems when, for example, one test mocks a function to allow a
certain series of function calls, and then a later test also uses that series of
function calls, but does not mock the function. In this case, if the first test
is run before the second test, they will both pass, but if the second test is
run first (or in isolation), it will fail because it calls an undefined
function. Test Mocks should not cause tests to fail if run in different order.

While it may be impossible to replicate a call to an undefined function, we can
simulate it by throwing an Exception. This change causes
`Handler::handle_function()` to throw an Exception if it doesn't find a matching
handler.